Organize the tornado extened classes.
- Create folder called tornado_ext
- Move iostream.py to tornado_ext
- Separate some of the logic from ProxyServer to tornado_ext.
- Update related clients and unittest.
